+++ src/client/screens/HomeScreen.tsx
@@ -9,7 +9,7 @@ import PaginationFooter from "../components/PaginationFooter";
 
 import { userProjectsFetch, userProjectsFetchPage } from "../actions/projects";
 import { userFetch } from "../actions/user";
-import { getJWT } from "../jwt";
+import { isUserLoggedIn } from "../jwt";
 import { State } from "../reducers";
 import { UserState } from "../reducers/user";
 import { Resource } from "../resource";
@@ -30,7 +30,7 @@ interface StateProps {
 }
 
 const HomeScreen = ({ projects, isSaving, duplicatedProject, user, pagination }: StateProps) => {
-  const isLoggedIn = getJWT() !== null;
+  const isLoggedIn = isUserLoggedIn();
   const projectList =
     "resource" in projects ? projects.resource.filter(project => !project.archived) : [];
 
@@ -125,7 +125,15 @@ const HomeScreen = ({ projects, isSaving, duplicatedProject, user, pagination }:
                 >
                   getting started guide
                 </Styled.a>{" "}
-                or see <Styled.a href="/maps">community maps</Styled.a>.
+                or see <Styled.a href="/maps">community maps</Styled.a>. If you want to start
+                mapping for an organization, you&rsquo;ll need to{" "}
+                <Styled.a
+                  href="https://github.com/PublicMapping/districtbuilder/wiki/Join-an-Organization"
+                  target="_blank"
+                >
+                  join the organization
+                </Styled.a>{" "}
+                first.
               </Text>
               <Flex>
                 <Styled.a